Added bookmarking in my work tab [#168962987]
This feature is for both students and teachers
