+++ src/clue/app-config.json
@@ -46,14 +46,24 @@
           "dataTestHeader": "my-work-section-workspaces",
           "dataTestItem": "my-work-list-items",
           "documentTypes": ["personal"],
-          "properties": ["!isDeleted"]
+          "properties": ["!isDeleted"],
+          "showStars": ["student", "teacher"]
         },
         {
           "title": "Problem Workspaces",
           "type": "problem-documents",
           "dataTestHeader": "my-work-section-investigations",
           "dataTestItem": "my-work-list-items",
-          "documentTypes": ["problem"]
+          "documentTypes": ["problem"],
+          "showStars": ["student", "teacher"]
+        },
+        {
+          "title": "Starred",
+          "type": "starred-problem-documents",
+          "dataTestHeader": "my-work-section-starred",
+          "documentTypes": ["personal", "problem"],
+          "properties": ["starred"],
+          "showStars": ["student", "teacher"]
         }
       ]
     },
@@ -76,7 +86,7 @@
           "dataTestHeader": "class-work-section-published",
           "dataTestItem": "class-work-list-items",
           "documentTypes": ["publication"],
-          "showStars": true
+          "showStars": ["teacher"]
         },
         {
           "className": "section learning-log published",
@@ -87,13 +97,13 @@
           "documentTypes": ["learningLogPublication"]
         },
         {
-          "className": "section problem published starred",
+          "className": "section problem starred",
           "title": "Starred",
           "type": "starred-problem-documents",
           "dataTestHeader": "class-work-section-starred",
           "documentTypes": ["publication"],
           "properties": ["starred"],
-          "showStars": true
+          "showStars": ["teacher"]
         }
       ]
     },
